commands: move generate under new command with new subcommand
